bot/requirements.txt
python-telegram-bot==5.3.0
psycopg2==2.7.1
retry==0.9.2
